Add files via upload
This is python code that includes a function for parsing EIS data formatted in 2 columns (Z' and Z'') from a web server and then plotting the data. The function can also be used to plot other 2 column data. Given below the function is an example of how to use it that pulls EIS data from the drops server.